Module libproc::libproc::kmesg_buffer

Get messages from the kernel message buffer

Functions

kmsgbuf

Get a message (String) from the kernel message ring buffer Turns out that reading to the end of an “infinite file” like “/dev/kmsg” with standard file reading methods will block at the end of file, so a workaround is required. Do the blocking reads on a thread that sends lines read back through a channel, and then return when the thread has blocked and can’t send anymore. Returning will end the thread and the channel.