Trait libproc::libproc::pid_rusage::PIDRUsage

pub trait PIDRUsage: Default {
    fn flavor() -> PidRUsageFlavor;

    fn memory_used(&self) -> u64;

    fn set_memory_used(&mut self, used: u64);
}

The PIDRUsage trait is needed for polymorphism on pidrusage types, also abstracting flavor in order to provide type-guaranteed flavor correctness

Required methods

fn flavor() -> PidRUsageFlavor

Return the PidRUsageFlavor for the implementing struct

fn memory_used(&self) -> u64

Memory used in bytes

fn set_memory_used(&mut self, used: u64)

Memory used in bytes