Module libproc::libproc::proc_pid

Get basic information about processes by PID

Structs

ListThreads

Struct for List of Threads

Enums

PidInfo

The PidInfo enum contains a piece of information about a processes

PidInfoFlavor

An enum used to specify what type of information about a process is referenced See http://opensource.apple.com/source/xnu/xnu-1504.7.4/bsd/kern/proc_info.c

ProcType

The ProcType type. Used to specify what type of processes you are interested in in other calls, such as listpids.

Traits

ListPIDInfo

The ListPIDInfo trait is needed for polymorphism on listpidinfo types, also abstracting flavor in order to provide type-guaranteed flavor correctness

PIDInfo

The PIDInfo trait is needed for polymorphism on pidinfo types, also abstracting flavor in order to provide type-guaranteed flavor correctness

Functions

am_root

Determine if the current user ID of this process is root

cwdself

Gets path of current working directory for the current process.

libversion

Get the major and minor version numbers of the native libproc library (Mac OS X)

listpidinfo

Get information on all running processes.

listpidsDeprecated

Returns the PIDs of the active processes that match the ProcType passed in

listpidspathDeprecated

Search through the current processes looking for open file references which match a specified path or volume.

name

Get the name of a process, using it’s process id (pid)

pidcwd

Gets the path of current working directory for the process with the provided pid.

pidinfo

Get info about a process, task, thread or work queue by specifying the appropriate type for T:

pidpath

Get the path of the executable file being run for a process

regionfilename

Get the filename associated with a memory region