Enum libproc::libproc::proc_pid::PidInfo

pub enum PidInfo {
    ListFDs(Vec<i32>),
    TaskAllInfo(TaskAllInfo),
    TBSDInfo(BSDInfo),
    TaskInfo(TaskInfo),
    ThreadInfo(ThreadInfo),
    ListThreads(Vec<i32>),
    RegionInfo(String),
    RegionPathInfo(String),
    VNodePathInfo(String),
    ThreadPathInfo(String),
    PathInfo(String),
    WorkQueueInfo(WorkQueueInfo),
}

The PidInfo enum contains a piece of information about a processes

Variants

ListFDs(Vec<i32>)

File Descriptors used by Process

TaskAllInfo(TaskAllInfo)

TBSDInfo(BSDInfo)

TaskInfo(TaskInfo)

ThreadInfo(ThreadInfo)

ListThreads(Vec<i32>)

A list of Thread IDs

RegionInfo(String)

RegionPathInfo(String)

VNodePathInfo(String)

ThreadPathInfo(String)

PathInfo(String)

The path of the executable being run as the process

WorkQueueInfo(WorkQueueInfo)