Trait libproc::libproc::proc_pid::ListPIDInfo

pub trait ListPIDInfo {
    type Item;

    // Required method
    fn flavor() -> PidInfoFlavor;
}

The ListPIDInfo trait is needed for polymorphism on listpidinfo types, also abstracting flavor in order to provide type-guaranteed flavor correctness

Required Associated Types

type Item

Item

Required Methods

fn flavor() -> PidInfoFlavor

Return the PidInfoFlavor of the implementing struct

Object Safety

This trait is not object safe.