Function libproc::processes::pids_by_path

pub fn pids_by_path(
    path: &Path,
    is_volume: bool,
    exclude_event_only: bool
) -> Result<Vec<u32>>

Returns the PIDs of active processes that reference reference an open file with the given path or volume, with or without files opened with the O_EVTONLY flag.

(Files opened with the O_EVTONLY flag will not prevent a volume from being unmounted).

Examples

Get the list of all running process IDs that have a specific filename open:

use std::path::Path;
use std::io::Write;
use libproc::processes;

let path = Path::new("/etc/hosts");
if let Ok(pids) = processes::pids_by_path(&path, false, false) {
    println!("Found {} processes accessing {}", pids.len(), path.display());
}

List all processes that have a file open on a specific volume; the path argument is used to get the filesystem device ID, and processes that have a file open on that same device ID match the filter:

use std::path::Path;
use std::io::Write;
use libproc::processes;

let path = Path::new("/Volumes/MountedDrive");
if let Ok(pids) = processes::pids_by_path(&path, true, false) {
    println!("Found {} processes accessing files on {}", pids.len(), path.display());
}