Function libproc::processes::pids_by_type_and_path

pub fn pids_by_type_and_path(
    filter: ProcFilter,
    path: &Path,
    is_volume: bool,
    exclude_event_only: bool
) -> Result<Vec<u32>>

Returns a filtered list of PIDs of active processes that reference reference an open file with the given path or volume, with or without files opened with the O_EVTONLY flag. Use a ProcFilter member to specify how to filter the list of PIDs returned.

(Files opened with the O_EVTONLY flag will not prevent a volume from being unmounted).

Examples

Get the list of process ids for child processes that have a specific filename open:

use std::path::Path;
use std::process;
use std::io::Write;
use libproc::processes;

let path = Path::new("/etc/hosts");
let filter = processes::ProcFilter::ByParentProcess { ppid: process::id() };
if let Ok(pids) = processes::pids_by_type_and_path(filter, &path, false, false) {
    println!("Found {} processes accessing {}", pids.len(), path.display());
}

List all processes within the current process group that have a file open on a specific volume; the path argument is used to get the filesystem device ID, and processes that have a file open on that same device ID match the filter:

use std::path::Path;
use std::process;
use std::io::Write;
use libproc::processes;

let path = Path::new("/Volumes/MountedDrive");
let filter = processes::ProcFilter::ByParentProcess { ppid: process::id() };
if let Ok(pids) = processes::pids_by_type_and_path(filter, &path, true, false) {
    println!("Found {} processes accessing files on {}", pids.len(), path.display());
}