Module libproc::libproc

libproc module provides library methods for getting various types of information about running processes.

In the future there will probably be a shorter list of public modules implementing a cleaner and reduced API - built on these methods.

Modules

kmesg_buffer

Get messages from the kernel message buffer

pid_rusage

Information about Process Resource Usage - added in Mac OS X 10.9

proc_pid

Get basic information about processes by PID

work_queue_info

Information about Work Queues - very MacOS specific