Module libproc::libproc

libproc module provides library methods for getting various types of information about running processes.

In the future there will probably be a shorter list of public modules implementing a cleaner and reduced API - built on these methods.

Modules

bsd_info

BSD specific information - very MacOS specific

kmesg_buffer

Get messages from the kernel message buffer

pid_rusage

Information about Process Resource Usage - added in Mac OS X 10.9

proc_pid

Get basic information about processes by PID

task_info

Information about Tasks - very MacOS specific

thread_info

Information about Threads running inside processes

work_queue_info

Information about Work Queues - very MacOS specific