Module libproc::libproc::proc_pid

Get basic information about processes by PID

Structs

ListThreads

Enums

PidInfo
The PidInfo enum contains a piece of information about a processes

PidInfoFlavor
An enum used to specify what type of information about a process is referenced

ProcType
The ProcType type. Used to specify what type of processes you are interested in in other calls, such as listpids.

Traits

ListPIDInfo

PIDInfo

Functions

am_root
Determine if the current user ID of this process is root

cwdself
Gets path of current working directory for the current process.

libversion
Get the major and minor version numbers of the native libproc library (Mac OS X)

listpidinfo
Get information on all running processes.

listpids
Returns the PIDs of the active processes that match the ProcType passed in

listpidspath

name
Get the name of a process

pidcwd
Gets the path of current working directory for the process with the provided pid.

pidinfo
Get info about a process

pidpath
Get the path of the executable file being run for a process

regionfilename
Get the filename associated with a memory region