Module libproc::libproc::proc_pid

Get basic information about processes by PID

Structs

ListThreads

Struct for List of Threads

Enums

PidInfo

The PidInfo enum contains a piece of information about a processes

PidInfoFlavor

An enum used to specify what type of information about a process is referenced See http://opensource.apple.com/source/xnu/xnu-1504.7.4/bsd/kern/proc_info.c

ProcType

The ProcType type. Used to specify what type of processes you are interested in in other calls, such as listpids.

Traits

ListPIDInfo

The ListPIDInfo trait is needed for polymorphism on listpidinfo types, also abstracting flavor in order to provide type-guaranteed flavor correctness

PIDInfo

The PIDInfo trait is needed for polymorphism on pidinfo types, also abstracting flavor in order to provide type-guaranteed flavor correctness

Functions

am_root

Return true if the calling process is being run by the root user, false otherwise

cwdself

Gets path of current working directory for the current process.

libversion

Get the major and minor version numbers of the native libproc library (Mac OS X)

listpidinfo

listpidinfo is not implemented on Linux - Pull Requests welcome - TODO

listpidsDeprecated

Returns the PIDs of the active processes that match the ProcType passed in

name

Get the name of a process, using it’s process id (pid)

pidcwd

Gets the path of current working directory for the process with the provided pid.

pidinfo

pidinfo not implemented on linux - Pull Requests welcome - TODO

pidpath

Get the path of the executable file being run for a process

regionfilename

Get the filename associated with a memory region