Trait libproc::libproc::proc_pid::ListPIDInfo

pub trait ListPIDInfo {
    type Item;
    fn flavor() -> PidInfoFlavor;
}

The ListPIDInfo trait is needed for polymorphism on listpidinfo types, also abstracting flavor in order to provide type-guaranteed flavor correctness

Associated Types

type Item

Item

Required methods

fn flavor() -> PidInfoFlavor

Return the PidInfoFlavor of the implementing struct